requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
click==7.1.2
cryptography==42.0.2
dnacentersdk==2.0.2
fastjsonschema==2.19.1
Flask==1.1.2
frozenlist==1.4.1
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
meraki==1.4.3
multidict==6.0.4
netmiko==4.3.0
ntc_templates==4.2.0
paramiko==3.4.0
pycparser==2.21
PyJWT==1.7.1
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==0.14.0
PyYAML==6.0.1
requests==2.24.0
requests-toolbelt==0.9.1
scp==0.14.5
six==1.16.0
textfsm==1.1.3
urllib3==1.25.10
webexteamssdk==1.6
Werkzeug==1.0.1
yarl==1.9.4
